update to latest SDK and themed layout
