+++ package.json
@@ -1,9 +1,9 @@
 {
-  "name": "zk-battleship",
-  "version": "0.1.0",
+  "name": "lab-l8",
+  "version": "0.0.1",
   "private": true,
   "dependencies": {
-    "@babbage/sdk-ts": "^0.1.11",
+    "@babbage/sdk-ts": "^0.2.49",
     "@emotion/react": "^11.11.3",
     "@emotion/styled": "^11.11.0",
     "@mui/material": "^5.15.10",
@@ -15,11 +15,13 @@
     "@types/react": "^18.0.28",
     "@types/react-dom": "^18.2.0",
     "babbage-scrypt-helpers": "^1.0.5",
+    "metanet-react-prompt": "^0.1.28",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-scripts": "5.0.1",
     "scrypt-ts": "^1.3.25",
     "typescript": "^4.8.4",
+    "use-async-effect": "^2.2.7",
     "whatsonchain": "^0.3.0"
   },
   "scripts": {
@@ -51,6 +53,7 @@
   "devDependencies": {
     "node-polyfill-webpack-plugin": "^2.0.1",
     "react-app-rewired": "^2.2.1",
+    "sass": "^1.77.8",
     "scrypt-cli": "^0.1.71"
   }
 }
