+++ public/manifest.json
@@ -1,5 +1,5 @@
 {
-  "short_name": "React App",
+  "short_name": "Locksmith React App",
   "name": "Create React App Sample",
   "icons": [
     {
