+++ src/App.scss
@@ -1,0 +1,3 @@
+a {
+  color:lightseagreen;
+}
