--- src/contracts/Locksmith.json
@@ -1,52 +1,0 @@
-{
-    "version": 9,
-    "compilerVersion": "1.19.2+commit.c3a20a0",
-    "contract": "Locksmith",
-    "md5": "d02e56a7db7f9f0c4845e9332160e4de",
-    "structs": [],
-    "library": [],
-    "alias": [],
-    "abi": [
-        {
-            "type": "function",
-            "name": "unlock",
-            "index": 0,
-            "params": [
-                {
-                    "name": "sig",
-                    "type": "Sig"
-                },
-                {
-                    "name": "pubKey",
-                    "type": "PubKey"
-                },
-                {
-                    "name": "__scrypt_ts_txPreimage",
-                    "type": "SigHashPreimage"
-                }
-            ]
-        },
-        {
-            "type": "constructor",
-            "params": [
-                {
-                    "name": "address",
-                    "type": "Ripemd160"
-                },
-                {
-                    "name": "lockUntilHeight",
-                    "type": "int"
-                },
-                {
-                    "name": "message",
-                    "type": "bytes"
-                }
-            ]
-        }
-    ],
-    "stateProps": [],
-    "buildType": "debug",
-    "file": "file:///D:/Kz/p2ppsr/repos/ws/babbage-scrypt-locking/artifacts/Locksmith.scrypt",
-    "hex": "2097dfd76851bf465e8f715593b217714858bbe9570ff3bd5e33840a34e20ff0262102ba79df5f8ae7604a9830f03c7933028186aede0675a16f025dc4f8be8eec0382201008ce7480da41702918d1ec8e6849ba32b4d65b1e40dc669c31a1e6306b266c0000000000<address><lockUntilHeight><message>615179040065cd1d9f695279587a75577a577a577a577a577a577a577a5179577a75567a567a567a567a567a567a0079567a75557a557a557a557a557a75757561587901c26151795a795a79210ac407f0e4bd44bfc207355a778b046225a7068fc59ee7eda43ad905aadbffc800206c266b30e6a1319c66dc401e5bd6b432ba49688eecd118297041da8074ce08105c795679615679aa0079610079517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e01007e81517a75615779567956795679567961537956795479577995939521414136d08c5ed2bf3ba048afe6dcaebafeffffffffffffffffffffffffffffff00517951796151795179970079009f63007952799367007968517a75517a75517a7561527a75517a517951795296a0630079527994527a75517a6853798277527982775379012080517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f517f7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e7c7e01205279947f7754537993527993013051797e527e54797e58797e527e53797e52797e57797e0079517a75517a75517a75517a75517a75517a75517a75517a75517a75517a75517a75517a75517a756100795779ac517a75517a75517a75517a75517a75517a75517a75517a75517a7561517a75517a756169587961007961007982775179517954947f75517958947f77517a75517a756161007901007e81517a7561517a7561527a75517a58796100796100798277517951790128947f755179012c947f77517a75517a756161007901007e81517a7561517a7561517a755179040065cd1d9f69007905feffffff009c6951795479a2695979a9557987695a795a79ac7777777777777777777777",
-    "sourceMapFile": ""
-}
